fixes -o (output logs), bin, and module usage
requiring this as a module did nothing because there was no way to
specify args, that is fixed now.

node modules with a command line interface should have a bin field in
the package.json file so node can create a shortcut in the
node_modules/.bin directory for end user ease of use.  That happens now.

when specifying -o option on the command line node-jsdoc-toolkit would
not output a log file, it would throw an error. Fixing this required a
shim for java.io.FileWriter class whose instance would be returned from
IO.open. That is done and the output log works now.